pyproject.toml
[tool.poetry]
name = "technician-booking-backend"
version = "0.1.0"
description = "Technician Booking System Backend with NLP Integration"
authors = ["Ericson Willians <ericsonwillians@protonmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
transformers = "^4.36.2"
torch = "^2.1.1"
datasets = "^2.16.1"
sentence-transformers = "^2.2.2"
accelerate = "^0.25.0"
python-dateutil = "^2.8.2"
typer = "^0.9.0"
rich = "^13.9.4"
pydantic-settings = "^2.7.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
mypy = "^1.6.1"
black = "^23.10.1"
flake8 = "^6.1.0"
isort = "^5.12.0"
httpx = "^0.25.0"

[tool.poetry.scripts]
booking-cli = "app.core.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"